% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dftest}
\alias{dftest}
\title{dftest a example for Rcpp}
\usage{
dftest(df, vname)
}
\arguments{
\item{df}{a data frame}

\item{vanme}{name of the col you wanna extract}
}
\value{
a vector which is the col you select
}
\description{
it is a example for Rcpp and devtools, which extract a column of a dataframe
}
\examples{
test<-data.frame(a = rep(1,5),b = rep(4,5))
aaa<-dftest(test,"a")
}
