% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMSAnnotate.R
\name{DMSAnnotate}
\alias{DMSAnnotate}
\title{The function annotates the methylDiff/methylBase obj.}
\usage{
DMSAnnotate(object, grAnno)
}
\arguments{
\item{object:}{A methylDiff/methyBase object}

\item{grAnno:}{A GRanges object}
}
\value{
A dataframe which contains site location and annotation. In the dataframe, there won't be the metainfo like p-values etc.
}
\description{
The function annotates the methylDiff/methylBase obj.
}
\examples{
data(ExampleAnnotate)
result<-DMSAnnotate()



}
