% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/employee.R
\name{employee}
\alias{employee}
\title{Generate a employee object.}
\usage{
employee(name = "Joe", salary = 5000, union = T,
  info = data.frame(x = rep(1, 10), y = rep(2, 10)))
}
\arguments{
\item{name}{the name, defalut is "Joe"}

\item{salary}{the salary, default is 5000}

\item{union}{boolean whether the man is in Union, defualt is T}
}
\description{
Generate a employee object.
}
\examples{
Wht<-employee(name="Wht",salary = 6000)
Wht
}
