% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findDMVS.R
\name{findDMVS}
\alias{findDMVS}
\title{The function find the DMVs defined by Xie et al in Cell 2013.
To identify each DMVs in a cell type, the genome was first divided in 1kb
bins and the DNA methylation level was averaged within each bin. Then a sliding
5kb window (with 1kb step) was used to identify regions that have an averaged
methylation level less than 0.15 in a 5kb window. Continuous regions resulting
from this analysis were then merged to form DMVs.}
\usage{
findDMVS(methyMatrix)
}
\arguments{
\item{methyMatrix:}{A matrix contains the methylation ratios. Columns are
samples and rows are sites. Each rowname is like 'chrX.NNNNN'}
}
\value{
A dataframe shows the methylation of 5kb-1kbstep regions of each sample.
}
\description{
The function find the DMVs defined by Xie et al in Cell 2013.
To identify each DMVs in a cell type, the genome was first divided in 1kb
bins and the DNA methylation level was averaged within each bin. Then a sliding
5kb window (with 1kb step) was used to identify regions that have an averaged
methylation level less than 0.15 in a 5kb window. Continuous regions resulting
from this analysis were then merged to form DMVs.
}
\examples{



}
