% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMethyVsExpression.R
\name{getMethyVsExpression}
\alias{getMethyVsExpression}
\title{The function get the correlation between methylation and expression}
\usage{
getMethyVsExpression(df1, df2)
}
\arguments{
\item{df1:}{A dataframe which contains the methylation,1st column is gene Symbol(or refseq name, gene id etc)}

\item{df2:}{A dataframe which contains the expression,1st column is gene Symbol(or refseq name, gene id etc)}
}
\value{
A dataframe which contains the p-value & corr coefficients
}
\description{
The function get the correlation between methylation and expression
}
\examples{
data("ExampleCorr")
result<-getMethyVsExpression(exMethyMatrix,exExpMatrix)



}
