% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetnAge-class.R
\docType{methods}
\name{makeMetnAge}
\alias{makeMetnAge}
\title{This function create the MetnAge object}
\usage{
makeMetnAge(metainfo, sampleinfo)
}
\arguments{
\item{metainfo}{a dataframe contains the Methylation perc, the
dataframe can be got from the \code{methylKit::percMethylation} -> GRanges
The first 5 cols should be chr,seqnames,start,end,width.
In fact the dataframe is construct from GRanges}

\item{sampleinfo}{a dataframe contain the sampleinfo. It must include ages.}
}
\value{
a Metnage object
}
\description{
If you want to create one MetnAge, you should use the function
named makeMetnAge.
}
\examples{
data("ExampleMethnAge")
testobj<-makeMetnAge(test,sampleInfo)

}
