% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetnAge-class.R
\docType{class}
\name{MetnAge-class}
\alias{MetnAge-class}
\title{This is a stupid DataStructure that storage the sample infomation
and the Methylation perc data}
\description{
If you want to create one MetnAge, you should use the function
named makeMetnAge.
}
\section{Slots}{

\describe{
\item{\code{MetaInfo}}{The methylation status for each site}

\item{\code{SampleInfo}}{The sample info always include age/subid/gender ...}

\item{\code{locInfo}}{The location infomation of each site (chr, bp, strand)}
}}

\author{
Wang Haotian (ritianjiang@gmx.com)
}
