% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCpghore.R
\name{plotCpghore}
\alias{plotCpghore}
\title{The function deals the cpg methylation status}
\usage{
plotCpghore(object, cpgFile, win_len, ...)
}
\arguments{
\item{object:}{A GRanges object that is the methylation perc file}

\item{grAnno:}{A GRanges object that is the cpgi&shores file}

\item{plot:}{... bool, plot or not (default: F)}

\item{len:}{... int, The length of cpgi you wanna normalized (default: 1000)}

\item{win_len:}{The slide window (default:10)}
}
\value{
A dataframe which contains cpgi & shore methylation status. the
}
\description{
The function deals the cpg methylation status
}
\examples{
data(ExampleCpgPlot)
result<-plotCpghore(testMeatrix,testCPG)



}
