% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGenic.R
\name{plotGenic}
\alias{plotGenic}
\title{Plot methylation status across genic regions(or some other features) and their up/down stream.}
\usage{
plotGenic(object, grAnno, numbin, wingbin, upstream = 2000,
  downstream = 500, scoreColumnName, Strand = FALSE)
}
\arguments{
\item{object:}{A GRanges object contains the}

\item{grAnno:}{A TxDb object containing the features. Please provide the strand info if you are ploting}

\item{numbin:}{bin number to split the feature regions we wanna plot}

\item{wingbin:}{bin number of the up/down stream}

\item{upstream:}{numeric; upstream bp to plot; default:2000}

\item{downstream:}{numeric; downstream bp to plot; default:500}

\item{scoreColumnName:}{character; The name of score in object.}

\item{Strand:}{boolean; Wheather use the strand info of object. default: F.}
}
\value{
A dataframe which contains site location and annotation. In the dataframe, there won't be the metainfo like p-values etc.
}
\description{
Plot methylation status across genic regions(or some other features) and their up/down stream.
}
\examples{
data(ExampleAnnotate)
result<-DMSAnnotate(methylDiff.obj, ExampleGencode)



}
