% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findemQTL.R
\docType{methods}
\name{findemQTL}
\alias{findemQTL}
\title{findemQTL}
\usage{
findemQTL(met, exp, cors = 1, siteFile, qFlag = F)
}
\arguments{
\item{met}{a data.frame contains methylation percentation}

\item{exp}{a data.frame contains gene and fpkm(or some others)
The 1st column is the gene names and the last is chrom}

\item{cors}{a numeric value of numbers of cores you wanna use. (default:1)}

\item{siteFile}{a dataframe contains the CpG site info}

\item{aFlag}{logical(default:F). Whether to use FDR correction.}
}
\value{
a dataframe contains site location info and potential regulated gene
and p-value(or qvalue in fact) and slope
}
\description{
This function find the potential emQTL by correlation test.
}
\examples{
data("ExampleemQTL")
res<-findemQTL(test[1:20,],chr5,cors = 4,siteFile = siteInfo)
res

}
